change to double return types

+++ src/ElectronScaleFactors.C
@@ -6,7 +6,7 @@
 class ElectronScaleFactors2016 {
 
   public:
-    static float LookupHeepSF(const float scEta) {
+    static double LookupHeepSF(const double scEta) {
       if(fabs(scEta)<1.566)
         return 0.983;
       else if(fabs(scEta)<2.5)
@@ -15,7 +15,7 @@ class ElectronScaleFactors2016 {
     }
 
     // this is the prompt uncertainty
-    static float LookupHeepSFSyst(const float scEta, const float et) {
+    static double LookupHeepSFSyst(const double scEta, const double et) {
       if(fabs(scEta)<1.566) {
         if(et < 90)
           return 0.01;
@@ -37,9 +37,9 @@ class ElectronScaleFactors2016 {
 class ElectronScaleFactors2017 {
 
   public:
-    static constexpr float zVtxSF = 0.991;
+    static constexpr double zVtxSF = 0.991;
 
-    static float LookupHeepSF(const float scEta)
+    static double LookupHeepSF(const double scEta)
     {
       if(fabs(scEta)<1.566)
         return 0.968;  // from the slides: https://indico.cern.ch/event/831669/contributions/3485543/attachments/1871797/3084930/ApprovalSlides_EE_v3.pdf
@@ -48,7 +48,7 @@ class ElectronScaleFactors2017 {
       return -1.;
     }
 
-    static float LookupHeepSFSyst(const float scEta, const float et) {
+    static double LookupHeepSFSyst(const double scEta, const double et) {
       if(fabs(scEta)<1.566) {
         if(et < 90)
           return 0.01;
@@ -69,7 +69,7 @@ class ElectronScaleFactors2017 {
 class ElectronScaleFactors2018 {
   public:
     // for HEEPv7.0-2018Prompt 
-    static float LookupHeepSF(const float scEta)
+    static double LookupHeepSF(const double scEta)
     {
       if(fabs(scEta)<1.566)
         return 0.969;
@@ -78,14 +78,14 @@ class ElectronScaleFactors2018 {
       return -1.;
     }
 
-    static float LookupHeepSFSyst(const float scEta, const float et) {
+    static double LookupHeepSFSyst(const double scEta, const double et) {
       return ElectronScaleFactors2017::LookupHeepSFSyst(scEta, et);
     }
 };
 
 class ElectronScaleFactorsRunII {
   public:
-    static float LookupHeepSF(const float scEta, const int analysisYear) {
+    static double LookupHeepSF(const double scEta, const int analysisYear) {
       if(analysisYear==2016)
         return ElectronScaleFactors2016::LookupHeepSF(scEta);
       else if(analysisYear==2017)
@@ -96,7 +96,7 @@ class ElectronScaleFactorsRunII {
         return -1;
     }
 
-    static float LookupHeepSFSyst(const float scEta, const float et, const int analysisYear) {
+    static double LookupHeepSFSyst(const double scEta, const double et, const int analysisYear) {
       if(analysisYear==2016)
         return ElectronScaleFactors2016::LookupHeepSFSyst(scEta, et);
       else if(analysisYear==2017)
